+++ notification-core/src/main/java/com/liveklass/notification/domain/vo/IdempotencyKey.java
@@ -1,25 +1,17 @@
-package com.liveklass.notification.domain;
+package com.liveklass.notification.domain.vo;
 
-import com.liveklass.common.error.ExceptionCreator;
 import com.liveklass.common.event.ChannelType;
 import com.liveklass.common.event.Topic;
-import com.liveklass.notification.domain.exception.OutboxException;
 
 import java.util.Objects;
 
-public record DedupKey(String value) {
+public record IdempotencyKey(String value) {
 
-    private static final int MAX_LENGTH = 512;
-
-    public DedupKey {
+    public IdempotencyKey {
         Objects.requireNonNull(value, "value must not be null");
-        if (value.length() > MAX_LENGTH) {
-            throw ExceptionCreator.create(OutboxException.INVALID_DEDUP_KEY,
-                    "length: " + value.length() + ", max: " + MAX_LENGTH);
-        }
     }
 
-    public static DedupKey of(
+    public static IdempotencyKey of(
             final Topic topic,
             final Long recipientId,
             final ChannelType channelType,
@@ -29,6 +21,6 @@ public record DedupKey(String value) {
         Objects.requireNonNull(recipientId, "recipientId must not be null");
         Objects.requireNonNull(channelType, "channelType must not be null");
         Objects.requireNonNull(referenceId, "referenceId must not be null");
-        return new DedupKey(topic.name() + ":" + recipientId + ":" + channelType.name() + ":" + referenceId);
+        return new IdempotencyKey(topic.name() + ":" + recipientId + ":" + channelType.name() + ":" + referenceId);
     }
 }
